+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+## 0.15.0-alpha.2 (2017-10-19)
+
 * Icon: no longer show a `pointer` cursor when hovering, when no `onClick` prop is given.
 * Checkbox: add support for `mdl` theme, including `skipTheme` prop.
 * Export `HoverableProps`, `PublicHoverableProps` flow types.